requirements.txt
bitsandbytes>=0.39.0
transformers
torch
accelerate>=0.26.0
google-genai
textstat
gradio